+++ Tests/SwiftLintFrameworkTests/RulesTests.swift
@@ -89,6 +89,38 @@ class RulesTests: XCTestCase {
                    testMultiByteOffsets: false)
     }
 
+    func testFirebaseConfigActivate() {
+        verifyRule(FirebaseConfigActivateRule.description)
+    }
+
+    func testFirebaseConfigDefaults() {
+        verifyRule(FirebaseConfigDefaultsRule.description)
+    }
+
+    func testFirebaseConfigFetch() {
+        verifyRule(FirebaseConfigFetchRule.description)
+    }
+
+    func testFirebaseCore() {
+        verifyRule(FirebaseCoreRule.description)
+    }
+
+    func testFirebaseDynamicLinksCustomSchemeURL() {
+        verifyRule(FirebaseDynamicLinksCustomSchemeURLRule.description)
+    }
+
+    func testFirebaseDynamicLinksUniversalLink() {
+        verifyRule(FirebaseDynamicLinksUniversalLinkRule.description)
+    }
+
+    func testFirebaseDynamicLinksSchemeURL() {
+        verifyRule(FirebaseDynamicLinksSchemeURLRule.description)
+    }
+
+    func testFirebaseInvites() {
+        verifyRule(FirebaseInvitesRule.description)
+    }
+
     func testFirstWhere() {
         verifyRule(FirstWhereRule.description)
     }
@@ -381,6 +413,14 @@ extension RulesTests {
             ("testExplicitTypeInterfaceRule", testExplicitTypeInterfaceRule),
             ("testFatalErrorMessageRule", testFatalErrorMessageRule),
             ("testFileLength", testFileLength),
+            ("testFirebaseConfigActivate", testFirebaseConfigActivate),
+            ("testFirebaseConfigDefaults", testFirebaseConfigDefaults),
+            ("testFirebaseConfigFetch", testFirebaseConfigFetch),
+            ("testFirebaseCore", testFirebaseCore),
+            ("testFirebaseDynamicLinksCustomSchemeURL", testFirebaseDynamicLinksCustomSchemeURL),
+            ("testFirebaseDynamicLinksUniversalLink", testFirebaseDynamicLinksUniversalLink),
+            ("testFirebaseDynamicLinksSchemeURL", testFirebaseDynamicLinksSchemeURL),
+            ("testFirebaseInvites", testFirebaseInvites),
             ("testFirstWhere", testFirstWhere),
             ("testForceCast", testForceCast),
             ("testForceTry", testForceTry),